services/message_service/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
boto3==1.35.99
requests==2.31.0
python-dotenv==1.0.1
